fix(Navigation): Make navigation clickable on mobile
Before the logo had a higher z-index then the Navigation component. On desktop this didn't matter
since the Logo component and the Navigation component didn't overlap each other. On mobile they did.
By giving the Navigation component a higher z-index and decreasing the z-index of the Logo component
this is solved

fix #1